requirements.txt
flask
requests
python-dotenv
flask-socketio   # <-- optional for RabbitMQ later
eventlet         # <-- optional for live updates
